Remove duplicated work processing hdr files in report [DEX-376] (#2268)
Remove duplication of work when processing each hdr file by supporting
more control over the output file formats. Now we allow the overall
latency distribution to be output in .hgrm but the latency over time to
be output in .csv instead of requiring both files to have the same
format.